pyproject.toml
[tool.poetry]
name = "postmatic"
version = "0.1.0"
description = ""
authors = ["Owez <root@ogriffiths.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
praw = "^7.2.0"
requests = "^2.25.1"
selenium = "^3.141.0"
Pillow = "^8.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
